Command split: Implement the "makefile" command in a separate class
