+++ yk
@@ -27,6 +27,7 @@ require_relative "lib/commands/reset_command"
 require_relative "lib/commands/restructure_command"
 require_relative "lib/commands/ybc_command"
 require_relative "lib/commands/ruby_command"
+require_relative "lib/commands/makefile_command"
 
 class CLI < Thor
 
@@ -110,7 +111,7 @@ class CLI < Thor
   def makefile(*module_names)
     modules = modules_from_names(module_names)
     with_modules modules do |mod|
-      mod.generate_makefiles
+      Commands::MakefileCommand.new(@config).apply(mod)
     end
   end
